share some station records
